Movde VT code to builtin
