+++ src/irr/asset/CGLSLTexturePackerBuiltinIncludeLoader.h
@@ -28,7 +28,7 @@ public:
     const char* getVirtualDirectoryName() const override { return "glsl/texture_packer/"; }
 
 private:
-	static std::string getDefinitions(const std::string& _path)
+	static std::string getUtilities(const std::string& _path)
 	{
 		auto args = parseArgumentsFromPath(_path.substr(_path.find_first_of('/') + 1, _path.npos));
 		if (args.size() < 4u)
@@ -73,15 +73,46 @@ private:
 			"\n#define TILE_PADDING " + args[ix_tile_padding] + "u" +
 			"\n#define PADDED_TILE_SIZE uint(PAGE_SZ+2*TILE_PADDING)" +
 			"\n\nconst vec2 packingOffsets[] = vec2[PAGE_SZ_LOG2+1]( vec2(0.0,0.0)," + tilePackingOffsetsStr() + ");" + 
-			"\n#endif\n";
+			R"(
+#define irr_glsl_WRAP_REPEAT 0u
+#define irr_glsl_WRAP_CLAMP 1u
+#define irr_glsl_WRAP_MIRROR 2u
+
+vec3 unpackPageID(in uint pageID)
+{
+	// this is optimal, don't touch
+	uvec2 pageXY = uvec2(pageID,pageID>>ADDR_Y_SHIFT)&uvec2(ADDR_X_MASK,ADDR_Y_MASK);
+	return vec3(vec2(pageXY),float(pageID>>ADDR_LAYER_SHIFT));
+}
+uvec2 unpackWrapModes(in uvec2 texData)
+{
+    return (texData>>uvec2(28u,30u)) & uvec2(0x03u);
+}
+uint unpackMaxMipInVT(in uvec2 texData)
+{
+    return (texData.y>>24)&0x0fu;
+}
+vec3 unpackVirtualUV(in uvec2 texData)
+{
+	// assert that PAGE_SZ_LOG2<8 , or change the line to uvec3(texData.yy<<uvec2(PAGE_SZ_LOG2,PAGE_SZ_LOG2-8u),texData.y>>16u)
+    uvec3 unnormCoords = uvec3(texData.y<<PAGE_SZ_LOG2,texData.yy>>uvec2(8u-PAGE_SZ_LOG2,16u))&uvec3(uvec2(0xffu)<<PAGE_SZ_LOG2,0xffu);
+    return vec3(unnormCoords);
+}
+vec2 unpackSize(in uvec2 texData)
+{
+	return vec2(texData.x&0xffffu,texData.x>>16u);
+}
+
+#endif
+)";
 	}
 
 protected:
 	core::vector<std::pair<std::regex, HandleFunc_t>> getBuiltinNamesToFunctionMapping() const override
 	{
 		return {
-			//definitions.glsl/addr_x_bits/addr_y_bits/pg_sz_log2/tile_padding
-			{ std::regex{"definitions\\.glsl/[0-9]+/[0-9]+/[0-9]+/[0-9]+"}, &getDefinitions },
+			//utils.glsl/addr_x_bits/addr_y_bits/pg_sz_log2/tile_padding
+			{ std::regex{"utils\\.glsl/[0-9]+/[0-9]+/[0-9]+/[0-9]+"}, &getUtilities },
 		};
 	}
 };
